pyproject.toml
[tool.poetry]
name = "find-projects"
version = "0.1.0"
description = ""
authors = ["ricklon <rianders@docs.rutgers.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
ollama = "^0.3.3"
python-dotenv = "^1.0.1"
tqdm = "^4.66.5"
aiofiles = "^24.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
